lib-common: fix iopc runtime compile error with GCC 10.
This commit is a partial cherry-pick of 9fabe2747a9f.

When running iopc with GCC 10, it was emitting some cast-function-type
errors in iop_hash.
This is fixed in 2019.1 with 9fabe2747a9f, but unfortunately, we cannot
cherry-pick the whole patch in 2016.4.

The solution is to only apply the modifications relative to the hash
functions.

+++ hash-iop.h
@@ -18,7 +18,7 @@
 
 struct iop_struct_t;
 
-typedef void (*iop_hash_f)(void *ctx, const void *input, int ilen);
+typedef void (*iop_hash_f)(void *ctx, const void *input, ssize_t ilen);
 
 enum {
     IOP_HASH_SKIP_MISSING = 1 << 0, /* Skip missing optional fields         */
